Add basic structure for skinning catch explosions
